# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AvailabilityResponse": {
        "properties": {
          "date": {
            "type": "string"
          },
          "partySize": {
            "type": "integer"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotResponse"
            },
            "type": "array"
          },
          "timezone": {
            "description": "IANA zone the `time` fields are expressed in.",
            "type": "string"
          }
        },
        "required": [
          "date",
          "partySize",
          "timezone",
          "slots"
        ],
        "type": "object"
      },
...
      "CreateReviewDto": {
        "properties": {
          "ambienceRating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "body": {
            "maxLength": 2000,
            "type": "string"
          },
          "foodRating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "reservationId": {
            "format": "uuid",
            "type": "string"
          },
          "serviceRating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "reservationId",
          "rating"
        ],
        "type": "object"
      },
...
      "ImageResponse": {
        "properties": {
          "height": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "is_cover": {
            "description": "The venue hero. Exactly one per owner.",
            "type": "boolean"
          },
          "position": {
            "type": "integer"
          },
          "urls": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Width in px → public URL. Pick the smallest that fits.",
            "example": {
              "160": "https://…/160.webp",
              "400": "…",
              "1200": "…"
            },
            "type": "object"
          },
          "width": {
            "description": "The ORIGINAL's width, for the aspect box.",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "urls",
          "width",
          "height",
          "position",
          "is_cover"
        ],
        "type": "object"
      },
...
      "OpeningHoursResponse": {
        "properties": {
          "closes_at": {
            "type": "string"
          },
          "day_of_week": {
            "description": "0=Sunday; null on a one-off date",
            "nullable": true,
            "type": "integer"
          },
          "name_ar": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "opens_at": {
            "description": "HH:MM on the restaurant's wall clock.",
            "type": "string"
          },
          "spans_midnight": {
            "type": "boolean"
          },
          "specific_date": {
            "description": "YYYY-MM-DD; null on a weekly row",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "name_en",
          "name_ar",
          "opens_at",
          "closes_at",
          "spans_midnight"
        ],
        "type": "object"
      },
...
      "ReportReviewDto": {
        "properties": {
          "note": {
            "maxLength": 1000,
            "type": "string"
          },
          "reason": {
            "description": "`not_my_visit` is the one that is about US rather than the reviewer — a review attached to the wrong reservation is a bug in the verified-diner guarantee, not a moderation question.",
            "enum": [
              "spam",
              "abusive",
              "not_my_visit",
              "wrong_venue",
              "other"
            ],
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "RestaurantProfileResponse": {
        "properties": {
          "address_ar": {
            "nullable": true,
            "type": "string"
          },
          "address_en": {
            "nullable": true,
            "type": "string"
          },
          "amenities": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "booking_mode": {
            "description": "instant | request",
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "cuisines": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "description_ar": {
            "nullable": true,
            "type": "string"
          },
          "description_en": {
            "nullable": true,
            "type": "string"
          },
          "hours": {
            "items": {
              "$ref": "#/components/schemas/OpeningHoursResponse"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "images": {
            "description": "The venue gallery, cover first then by position. Empty for a venue with no photos — which the client draws as a designed empty state, never as a broken image.",
            "items": {
              "$ref": "#/components/schemas/ImageResponse"
            },
            "type": "array"
          },
          "lat": {
            "nullable": true,
            "type": "number"
          },
          "lng": {
            "nullable": true,
            "type": "number"
          },
          "name_ar": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "neighborhood": {
            "nullable": true,
            "type": "string"
          },
          "phone": {
            "nullable": true,
            "type": "string"
          },
          "policies": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "price_band": {
            "nullable": true,
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "rating_count": {
            "type": "integer"
          },
          "slug": {
            "type": "string"
          },
          "timezone": {
            "description": "IANA zone every wall-clock time here is in.",
            "type": "string"
          },
          "website": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name_en",
          "name_ar",
          "cuisines",
          "city",
          "rating",
          "rating_count",
          "amenities",
          "images",
          "timezone",
          "booking_mode",
          "hours"
        ],
        "type": "object"
      },
      "RestaurantResponse": {
        "properties": {
          "city": {
            "type": "string"
          },
          "descriptionAr": {
            "nullable": true,
            "type": "string"
          },
          "descriptionEn": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "nameAr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "neighborhood": {
            "nullable": true,
            "type": "string"
          },
          "priceBand": {
            "nullable": true,
            "type": "integer"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "status",
          "nameEn",
          "nameAr",
          "city"
        ],
        "type": "object"
      },
      "ReviewPageResponse": {
        "properties": {
          "next_cursor": {
            "description": "Keyset, not offset — a new review must not shift page two.",
            "nullable": true,
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/ReviewSummaryResponse"
          }
        },
        "required": [
          "summary",
          "results"
        ],
        "type": "object"
      },
      "ReviewResponse": {
        "properties": {
          "ambience_rating": {
            "nullable": true,
            "type": "integer"
          },
          "author": {
            "description": "First name and a surname initial (\"Nour H.\"). Never the full name a diner gave at registration.",
            "example": "Nour H.",
            "type": "string"
          },
          "body": {
            "description": "Nullable — stars alone is a complete review.",
            "nullable": true,
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "food_rating": {
            "nullable": true,
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "owner_replied_at": {
            "nullable": true,
            "type": "string"
          },
          "owner_reply": {
            "nullable": true,
            "type": "string"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "service_rating": {
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "rating",
          "author",
          "created_at"
        ],
        "type": "object"
      },
      "ReviewSummaryResponse": {
        "properties": {
          "breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Star figure → how many gave it. Always all five keys.",
            "example": {
              "1": 3,
              "2": 5,
              "3": 16,
              "4": 78,
              "5": 210
            },
            "type": "object"
          },
          "rating": {
            "description": "Computed from the reviews themselves, not from the cached column.",
            "type": "number"
          },
          "rating_count": {
            "type": "integer"
          }
        },
        "required": [
          "rating",
          "rating_count",
          "breakdown"
        ],
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "availability_filtered": {
            "type": "boolean"
          },
          "estimated_total": {
            "type": "integer"
          },
          "next_cursor": {
            "nullable": true,
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "results",
          "estimated_total",
          "availability_filtered"
        ],
        "type": "object"
      },
      "SearchResultResponse": {
        "properties": {
          "cover": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageResponse"
              }
            ],
            "description": "The venue hero, or null. Fetched for the whole page in ONE query — a search list that asked per row would issue twenty requests over a Cairo mobile connection to render its first screenful.",
            "nullable": true
          },
          "cuisines": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "distance_km": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "name_ar": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "neighborhood": {
            "nullable": true,
            "type": "string"
          },
          "next_available": {
            "description": "Local HH:MM teasers. A HINT, not an offer — no absolute instant is given precisely so no client can treat one as bookable.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "price_band": {
            "nullable": true,
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "rating_count": {
            "type": "integer"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name_en",
          "name_ar",
          "cuisines",
          "rating",
          "rating_count"
        ],
        "type": "object"
      },
...
      "SlotResponse": {
        "properties": {
          "startsAt": {
            "description": "Absolute instant, ISO-8601 UTC. POST this back.",
            "type": "string"
          },
          "time": {
            "description": "HH:MM on the RESTAURANT'S wall clock.",
            "type": "string"
          },
          "zones": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "time",
          "startsAt",
          "zones"
        ],
        "type": "object"
      },
...
      "UpdateRestaurantDto": {
        "properties": {
          "bookingMode": {
            "enum": [
              "instant",
              "request"
            ],
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "cuisines": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "descriptionAr": {
            "type": "string"
          },
          "descriptionEn": {
            "type": "string"
          },
          "nameAr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "priceBand": {
            "type": "number"
          },
          "slotIntervalMin": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Restaurant reservation platform — Egypt/MENA",
    "title": "SAHRA API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/owner/restaurants/{id}": {
      "get": {
        "operationId": "OwnerRestaurantsController_getOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of my restaurants",
        "tags": [
          "owner:restaurants"
        ]
      },
      "patch": {
        "operationId": "OwnerRestaurantsController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRestaurantDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update profile, policies, amenities",
        "tags": [
          "owner:restaurants"
        ]
      }
    },
...
    "/v1/restaurants/search": {
      "get": {
        "operationId": "SearchController_find",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "example": "sequoia",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cuisine",
            "required": false,
            "schema": {
              "example": "egyptian",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "neighborhood",
            "required": false,
            "schema": {
              "example": "Zamalek",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "price_band",
            "required": false,
            "schema": {
              "example": 3,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "rating_min",
            "required": false,
            "schema": {
              "example": 4,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "example": 30.0622,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": false,
            "schema": {
              "example": 31.2185,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "example": 5,
              "type": "string"
            }
          },
          {
            "description": "YYYY-MM-DD; needs party_size",
            "in": "query",
            "name": "available_at",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "party_size",
            "required": false,
            "schema": {
              "example": 2,
              "type": "string"
            }
          },
          {
            "description": "comma-separated",
            "in": "query",
            "name": "amenities",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "enum": [
                "relevance",
                "rating",
                "distance"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "example": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Discovery search — text + facets, availability post-filtered",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/restaurants/{idOrSlug}": {
      "get": {
        "operationId": "PublicRestaurantsController_profile",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantProfileResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "restaurant_not_found — also returned for a venue that is not active"
          }
        },
        "summary": "Public restaurant profile, by id or slug",
        "tags": [
          "restaurants"
        ]
      }
    },
...
    "/v1/restaurants/{idOrSlug}/reviews": {
      "get": {
        "operationId": "RestaurantReviewsController_listReviews",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ISO instant from `next_cursor`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPageResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "restaurant_not_found"
          }
        },
        "summary": "A venue's published reviews, newest first",
        "tags": [
          "reviews"
        ]
      }
    },
    "/v1/restaurants/{id}/availability": {
      "get": {
        "operationId": "AvailabilityController_getSlots",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "example": "2026-08-02",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "party_size",
            "required": true,
            "schema": {
              "example": 2,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Bookable slots for a date and party size",
        "tags": [
          "availability"
        ]
      }
    },
    "/v1/reviews": {
      "post": {
        "operationId": "ReviewsController_createReview",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": ""
          },
          "201": {
            "description": "Posted, and live immediately"
          },
          "400": {
            "description": "validation_failed — including `photo_ids`, which is not accepted; see the DTO"
          },
          "403": {
            "description": "review_not_eligible (the visit did not happen) or review_too_early (the table time is not over)"
          },
          "404": {
            "description": "reservation_not_found — ALSO for another diner's reservation, deliberately"
          },
          "409": {
            "description": "review_already_exists — one review per visit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Review a visit that happened",
        "tags": [
          "reviews"
        ]
      }
    },
    "/v1/reviews/{id}/report": {
      "post": {
        "operationId": "ReviewsController_reportReview",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Already reported by this diner"
          },
          "201": {
            "description": "Reported"
          },
          "400": {
            "description": "cannot_report_own_review"
          },
          "404": {
            "description": "review_not_found — ALSO for a review that is not published, deliberately"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report a review (recorded; the queue is A-3)",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "servers": [],
...
}